for the work experience. i have been working as a junior developer at megaworld corporation from Oct 2024 until the present. during my stay there here are the list of my accomplishments:
- Developed data-driven dashboards with Google Charts and backend integration via Laravel, streamlining HR analytics for 3000+ employees.
- Engineered a real-time biometric system dashboard with Laravel and JavaScript to monitor uptime/downtime across 130+ site offices.
- Provided technical support during the transition from manual attendance systems to biometric systems with real-time log uploading across multiple offices.

for the education, i studied at the University of the Philippines Diliman with a degree in Bachelor of Science in Electronics Engineering with a grade of cum laude. my stay there was from aug 2019 - sept 2023.

for the projects, i have a coffee shop website, dental clinic website, and a word search puzzle solver

for hobbies, i enjoy traveling, watching movies, and playing video games

remove the "Designed with passion and coded with Next.js & Tailwind CSS."